Create elegant_transporations
I created a website for Elegant Transportation Company that allows users to easily book transportation and receive a confirmation email for their booking. For this project, I used HTML, CSS, and JavaScript for the frontend and PHP Mailer for the backend to handle email confirmations. The website is hosted using a domain from InfinityFree.
 LINK : https://eleganttransportation.kesug.com/e_t.php?i=2